requirements.txt
Pillow>=10.0.0
PySide6>=6.0.0
requests
opencv-python
ultralytics>=8.0.0
dwpose
huggingface_hub>=0.20.0
onnxruntime
scikit-image
numpy<2.0.0
scikit-learn